% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTitle.R
\name{addTitle}
\alias{addTitle}
\title{Add title to the leaflet}
\usage{
addTitle(object, text, color = "black", fontSize = "20px",
 fontFamily = "Sans", leftPosition = 50, topPosition = 2)
}
\arguments{
\item{object}{an leaflet object}

\item{text}{content of the title}

\item{color}{title color, default: 'black'}

\item{fontSize}{font size for the title default: '20px'}

\item{fontFamily}{font family for the title, default: 'Sans'}

\item{leftPosition}{the percent of the title postition to left, default: 50}

\item{topPosition}{the percent of the title postition to top, default:2}
}
\description{
Function for creating a h1 title to the leaflet
}
\examples{
demomap('china') \%>\% addTitle("Hello World")

}

