% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Amap.R
\name{amap}
\alias{amap}
\title{Load amap to leaflet}
\usage{
amap(map,attribution = '&copy; <a href="http://amap.com">amp.com</a >',...)
}
\arguments{
\item{map}{a leaflet object}

\item{attribution}{attribution of the map}

\item{...}{other paramter pass to the addTiles function}
}
\description{
Simple function like addTiles()
}
\examples{
if(require(leaflet)){
  leaflet() \%>\% amap()
}

}
