% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoMap.R
\name{demomap}
\alias{demomap}
\title{Show the basic shape of the data}
\usage{
demomap (mapName)
}
\arguments{
\item{mapName}{mapName for loading, eg. 'china', 'city', ...}
}
\description{
This function create a leaflet map from a name of the mapType like china, city
}
\examples{
demomap("china")
}

