% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{leafletcn.map.names}
\alias{leafletcn.map.names}
\alias{.triList}
\title{Demo dataset and other function}
\format{A data set cities and file names}
\description{
A data file of words and frequency from tm package
}
