% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafletGeo.R
\name{leafletGeo}
\alias{leafletGeo}
\title{Create a leaflet object from a data.frame}
\usage{
leafletGeo(mapName, dat = NULL, namevar = NULL, valuevar = NULL)
}
\arguments{
\item{mapName}{mapName for loading, eg. 'china', 'city', ...}

\item{dat}{a data.frame contain regions and values}

\item{namevar}{show which feature is chosen for name variable}

\item{valuevar}{show which featue is chosen for value variable}
}
\description{
Function for creating a leaflet object from a data.frame and
mapType
}
\examples{
dat = data.frame(regionNames("city"),
                              runif(384))
map = leafletGeo("china", dat)
map \%>\% addTiles() \%>\%
   leaflet::addPolygons(stroke = stroke,
   smoothFactor = smoothFactor,
   fillOpacity = fillOpacity,
   weight = weight,
   color = ~pal(value),
   popup = ~htmltools::htmlEscape(popup)
   ) \%>\%
     leaflet::addLegend("bottomright", pal = pal, values = ~value,
                          title = legendTitle,
                 labFormat = leaflet::labelFormat(prefix = ""),
                 opacity = 1)
}

