% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapNames}
\alias{.triList}
\alias{mapNames}
\title{Demo dataset and other function}
\format{A data set cities and file names}
\usage{
mapNames

.triList
}
\description{
A data file of words and frequency from tm package
}
\keyword{datasets}

