% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.geoShape.R
\name{read.geoShape}
\alias{read.geoShape}
\title{Read geoshape file into R}
\usage{
read.geoShape(txt)
}
\arguments{
\item{txt}{a JSON string, URL or file}
}
\description{
Function for reading geojson shape file into R, come out to be a SpatialPolygonsDataFrame object
}
\examples{
require(sp)
filePath = system.file("geojson/china.json",package = "leafletCN")
map = read.geoShape("inst/geojson/11.json")
plot(map)
}

