% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionNames.R
\name{regionNames}
\alias{regionNames}
\title{Show regions in submaps}
\usage{
regionNames(mapName)
}
\arguments{
\item{mapName}{for loading, eg. 'china', 'city', ...}
}
\description{
show regions in the map
}
\examples{
regionNames('city')
}
